% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_donde_se_gasta.R
\name{choose_donde_se_gasta}
\alias{choose_donde_se_gasta}
\title{Definir la seccion de '¿Dónde se gasta?'}
\usage{
choose_donde_se_gasta(x, departamento_meta = NULL)
}
\arguments{
\item{departamento_meta}{chr. Código de departamento.
Para obtener "Callao" usar "07", y para obtener 'Exterior' usar "98".}
}
\value{
Una lista con nombres correspondientes a los endpoints de la API
}
\description{
Definir la seccion de '¿Dónde se gasta?'
}
