% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consultar.R
\name{consultar}
\alias{consultar}
\title{Ejecutar consulta de seguimiento a la ejecución presupuestal}
\usage{
consultar(x, ...)
}
\arguments{
\item{x}{Objeto de clase 'sep_df' recipiente de los parámetros de la consulta deseada}
}
\value{
Un data.frame con clase \if{html}{\out{<transpaeco>}} que contiene el resultado de la consulta solicitada
}
\description{
Ejecutar consulta de seguimiento a la ejecución presupuestal
}
