% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_como_se_estructura.R
\name{elegir_como_se_estructura}
\alias{elegir_como_se_estructura}
\title{Definir la seccion de '¿Cómo se estructura el gasto?'}
\usage{
elegir_como_se_estructura(
  x,
  generica = NULL,
  subgenerica = NULL,
  detalle_subgenerica = NULL,
  especifica = NULL,
  detalle_especifica = NULL
)
}
\arguments{
\item{x}{Objeto de clase 'sep_df' recipiente de los parámetros de la consulta deseada}

\item{generica}{chr. Código de genérica de forma "X-X-X"}

\item{subgenerica}{chr. Código de sub-genérica de forma "X-X-X"}

\item{detalle_subgenerica}{chr. Código de detalle de sub-genérica de forma "X-X-X-X"}

\item{especifica}{chr. Código de específica de forma "X-X-X-X-X"}

\item{detalle_especifica}{chr. Código de detalle de específica de forma "X-X-X-X-X-X"}
}
\value{
Una lista con nombres correspondientes a los endpoints de la API
}
\description{
Estos endpoints se construyen secuencialmente en base a los valores de generica.
}
\details{
Para escribir el código de genérica se debe alterar un poco la representación típica.
Por ejemplo, la genérica "7-28" debe pasarse a \code{generica} como "7-2-8".

A partir del nivel \code{subgenerica} debe omitirse solo el primer dígito de la genérica.
Por ejemplo, la sub-genérica "7-28.2" debe pasarse como "2-8-2" y para el
detalle de la sub-genérica "7-28.2.1" debe pasarse "2-8-2-1". La específica y
detalle de específica siguen este patrón.
}
