% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_cuando_se_hizo_gasto.R
\name{elegir_cuando_se_hizo_gasto}
\alias{elegir_cuando_se_hizo_gasto}
\title{Definir la seccion de '¿Cuándo se hizo el gasto?'}
\usage{
elegir_cuando_se_hizo_gasto(x, trimestre = NULL, mes = NULL)
}
\arguments{
\item{x}{Objeto de clase 'sep_df' recipiente de los parámetros de la consulta deseada}

\item{trimestre}{int. Número del 1 al 4}

\item{mes}{int. Número del 1 al 12}
}
\value{
Una lista con nombres correspondientes a los endpoints de la API
}
\description{
Definir la seccion de '¿Cuándo se hizo el gasto?'
}
