% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elegir_destino.R
\name{elegir_destino}
\alias{elegir_destino}
\alias{elegir_en_que_se_gasta}
\title{Elegir destino de uso del presupuesto.}
\usage{
elegir_destino(
  x,
  categoria_presupuestal = NULL,
  producto = NULL,
  actividad = NULL,
  funcion = NULL,
  division_funcional = NULL,
  grupo_funcional = NULL,
  meta = NULL
)

elegir_en_que_se_gasta(
  x,
  categoria_presupuestal = NULL,
  producto = NULL,
  actividad = NULL,
  funcion = NULL,
  division_funcional = NULL,
  grupo_funcional = NULL,
  meta = NULL
)
}
\arguments{
\item{x}{Objeto de clase \verb{<transpaeco>} recipiente de los parámetros de la consulta deseada}

\item{categoria_presupuestal}{chr. Código de programa presupuestal de forma "0000" (4 digitos).
Para Acciones Centrales usar "9001" y para APnoP usar "9002".}

\item{producto}{chr. Código de producto o proyecto de inversión de forma "0000000" (7 digitos).}

\item{actividad}{chr. Código de actividad, acción de inversión u obra, de forma "0000000" (7 digitos).}

\item{funcion}{chr. Código de función de forma "00" (2 digitos).}

\item{division_funcional}{chr. Código de división funcional de forma "000" (3 digitos).}

\item{grupo_funcional}{chr. Código de grupo funcional de forma "0000" (4 digitos).}

\item{meta}{chr. Código de meta. No tiene forma estandarizada y no es obtenerla a partir de la información provista.
Es el parametro de consulta mas problematico.}
}
\value{
Un data.frame con clase \verb{<transpaeco>}
}
\description{
Estas funciones permiten desglosar la busqueda de acuerdo al destino
de uso del presupuesto. Debido a que solo se pueden usar en el módulo de gasto,
ambas tienen el mismo comportamiento.
}
\examples{
iniciar_transparencia_economica(modulo = "gasto") \%>\%
    elegir_periodo_anual(2022) \%>\%
    elegir_en_que_se_gasta(categoria_presupuestal = "todos")

iniciar_transparencia_economica(modulo = "gasto") \%>\%
    elegir_periodo_anual(2022) \%>\%
    elegir_en_que_se_gasta(categoria_presupuestal = "todos") \%>\%
    consultar()

}
