% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elegir_donde_se_gasta.R
\name{elegir_donde_se_gasta}
\alias{elegir_donde_se_gasta}
\title{Definir la seccion de '¿Dónde se gasta?'}
\usage{
elegir_donde_se_gasta(x, departamento_meta = NULL)
}
\arguments{
\item{x}{Objeto de clase 'sep_df' recipiente de los parámetros de la consulta deseada}

\item{departamento_meta}{chr. Código de departamento.
Para obtener "Callao" usar "07", y para obtener 'Exterior' usar "98".}
}
\value{
Una lista con nombres correspondientes a los endpoints de la API
}
\description{
Definir la seccion de '¿Dónde se gasta?'
}
