% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elegir_years.R
\name{elegir_years}
\alias{elegir_years}
\title{Elegir periodo anual de consulta}
\usage{
elegir_years(x, years)
}
\arguments{
\item{x}{Objeto de clase 'sep_df' recipiente de los parámetros de la consulta deseada}

\item{years}{numeric. Periodo anual de consulta}
}
\value{
sep_df
}
\description{
Elegir periodo anual de consulta
}
