% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasto.R
\name{gasto}
\alias{gasto}
\title{Consulta de ejecución presupuestal}
\usage{
gasto(year, ..., actualizacion = "diaria")
}
\arguments{
\item{year}{Año para el cual se quiere obtener la información a partir de 1999.}

\item{...}{Otros argumentos pasados a \code{make_gasto_query()}}

\item{actualizacion}{Puede ser "diaria" o "mensual"}
}
\value{
Un tibble con la información requerida
}
\description{
Consulta de ejecución presupuestal
}
