% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_query_list.R
\name{make_gasto_query}
\alias{make_gasto_query}
\title{Armar el query detallado para consulta de ejecución presupuestal}
\usage{
make_gasto_query(
  year = NULL,
  nivel = NULL,
  sector = NULL,
  pliego = NULL,
  unidad_ejecutora = NULL,
  goblocal_o_manc = NULL,
  mancomunidad = NULL,
  departamento = NULL,
  provincia = NULL,
  municipalidad = NULL,
  categoria_presupuestal = NULL,
  producto = NULL,
  actividad = NULL,
  funcion = NULL,
  division_funcional = NULL,
  grupo_funcional = NULL,
  meta = NULL,
  fuente_financiamiento = NULL,
  rubro = NULL,
  tipo_de_recurso = NULL,
  generica = NULL,
  subgenerica = NULL,
  detalle_subgenerica = NULL,
  especifica = NULL,
  detalle_especifica = NULL,
  departamento_meta = NULL,
  trimestre = NULL,
  mes = NULL
)
}
\arguments{
\item{year}{Año del cual se quiere obtener la información}

\item{nivel}{chr. Nivel de gobierno. Puede ser "E" (gobierno nacional),
"M" (gobiernos locales) o "R" (gobiernos regionales). Si es "E" o "R",
le corresponde \code{sector}. Si es "M", le corresponde \code{goblocal_o_manc}.}

\item{sector}{chr. Codigo de sector, de forma "XX" (2).}

\item{pliego}{chr. Código de pliego, de forma "XXX" (3).}

\item{unidad_ejecutora}{chr. Código de Unidad Ejecutora. No tiene forma estandar}

\item{goblocal_o_manc}{chr. Escoger entre "M" (gobierno local)
o "N" (mancomunidad). Si es "M", desbloque \code{departamento}, \code{provincia} y \code{municipalidad}.
Si es "N", le corresponde \code{mancomunidad}.}

\item{mancomunidad}{chr. Código de mancomunidad, de forma "XXXXXX" (6).}

\item{departamento}{chr. Código de departamento. El valor "07" corresponde a Callao.}

\item{provincia}{chr. Código de provincia. El valor "01" corresponde a capital de departamento.}

\item{municipalidad}{chr. Código de municipalidad, de forma "XXXXXX". No corresponde con ubigeo.}

\item{categoria_presupuestal}{chr. Código de programa presupuestal de forma "XXXX" (4).
Para Acciones Centrales usar "9001" y para APnoP usar "9002".}

\item{producto}{chr. Código de producto o proyecto de inversión de forma "XXXXXXX" (7).}

\item{actividad}{chr. Código de actividad, acción de inversión u obra, de forma "XXXXXXX" (7).}

\item{funcion}{chr. Código de función de forma "XX" (2).}

\item{division_funcional}{chr. Código de división funcional de forma "XXX" (3).}

\item{grupo_funcional}{chr. Código de grupo funcional de forma "XXXX" (4).}

\item{meta}{chr. Código de meta. No tiene forma estandarizada.}

\item{fuente_financiamiento}{int. Código de fuente de financiamiento.
En su versión más reciente va del 1 al 5.}

\item{rubro}{chr. código de rubro, de  forma "XX" (2).}

\item{tipo_de_recurso}{chr, Código de tipo de recurso.
No tiene forma estandar, puede ser de uno o dos caracteres, que pueden
ser letras o números.}

\item{generica}{chr. Código de genérica de forma "X-X-X"}

\item{subgenerica}{chr. Código de sub-genérica de forma "X-X-X"}

\item{detalle_subgenerica}{chr. Código de detalle de sub-genérica de forma "X-X-X-X"}

\item{especifica}{chr. Código de específica de forma "X-X-X-X-X"}

\item{detalle_especifica}{chr. Código de detalle de específica de forma "X-X-X-X-X-X"}

\item{departamento_meta}{chr. Código de departamento.
Para obtener "Callao" usar "07", y para obtener 'Exterior' usar "98".}

\item{trimestre}{int. Número del 1 al 4}

\item{mes}{int. Número del 1 al 12}
}
\value{
Una lista cuyos elementos tienen el key-value de los parámetros de la API
}
\description{
Armar el query detallado para consulta de ejecución presupuestal
}
