% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_gasto_query.R
\name{perform_gasto_query}
\alias{perform_gasto_query}
\title{Ejecutar consulta de transparencia economica}
\usage{
perform_gasto_query(gasto_query, actualizacion = "diaria")
}
\arguments{
\item{gasto_query}{Lista de parámetros de query provistos por \code{make_gasto_query()}}

\item{actualizacion}{chr. Uno de c("mensual", "diaria")}
}
\value{
Un tibble con los datos consultados
}
\description{
Ejecutar consulta de transparencia economica
}
\examples{
make_gasto_query(2022) |> perform_gasto_query()
}
