% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_html_body.R
\name{retrieve_html_body}
\alias{retrieve_html_body}
\title{Obtener el cuerpo de la respuesta HTML}
\usage{
retrieve_html_body(request)
}
\arguments{
\item{request}{Objeto de clase "httr2_request"}
}
\value{
Objeto de clase "html_node"
}
\description{
Esta operación toma la "httr2_request" definida, la realiza en el servidor remoto
y obtiene solo el cuerpo de la respuesta HTML. Ha sido definido así porque
sirve como paso intermedio para la obtención de la tabla.
}
