% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_response_tbl.R
\name{retrive_response_tbl}
\alias{retrive_response_tbl}
\title{Obtener la tabla de ejecucion de gasto}
\usage{
retrive_response_tbl(response)
}
\arguments{
\item{response}{Un objeto de clase "xml_node" retornado por retrieve_html_body()}
}
\value{
Un tibble
}
\description{
Obtener la tabla de ejecucion de gasto
}
