% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep_df.R
\name{seguimiento_ep}
\alias{seguimiento_ep}
\title{Iniciar una consulta de seguimiento de ejecución presupuestal}
\usage{
seguimiento_ep(actualizacion = "diaria")
}
\arguments{
\item{actualizacion}{chr. Uno de c("mensual", "diaria")}
}
\value{
Objeto con clase 'sep_df' en el que se pueden agregar parámetros de consulta
}
\description{
Iniciar una consulta de seguimiento de ejecución presupuestal
}
