% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cuando_se_hizo.R
\name{set_cuando_se_hizo_gasto}
\alias{set_cuando_se_hizo_gasto}
\title{Definir la seccion de '¿Cuándo se hizo el gasto?'}
\usage{
set_cuando_se_hizo_gasto(trimestre = NULL, mes = NULL)
}
\arguments{
\item{trimestre}{int. Número del 1 al 4}

\item{mes}{int. Número del 1 al 12}
}
\value{
Una lista con nombres correspondientes a los endpoints de la API
}
\description{
Definir la seccion de '¿Cuándo se hizo el gasto?'
}
