% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consultar.R
\name{consultar}
\alias{consultar}
\title{Ejecutar consulta de seguimiento a la ejecución presupuestal}
\usage{
consultar(x)
}
\arguments{
\item{x}{Objeto de clase 'sep_df' que contiene los parámetros de la consulta deseada}
}
\value{
sep_df. Tabla que contiene el resultado de la consulta solicitada
}
\description{
Ejecutar consulta de seguimiento a la ejecución presupuestal
}
