% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elegir_destino.R
\name{elegir_destino}
\alias{elegir_destino}
\alias{elegir_en_que_se_gasta}
\title{Elegir destino de uso del presupuesto.}
\usage{
elegir_destino(
  x,
  categoria_presupuestal = NULL,
  producto = NULL,
  actividad = NULL,
  funcion = NULL,
  division_funcional = NULL,
  grupo_funcional = NULL,
  meta = NULL
)

elegir_en_que_se_gasta(
  x,
  categoria_presupuestal = NULL,
  producto = NULL,
  actividad = NULL,
  funcion = NULL,
  division_funcional = NULL,
  grupo_funcional = NULL,
  meta = NULL
)
}
\arguments{
\item{x}{Objeto de clase 'sep_df' recipiente de los parámetros de la consulta deseada}

\item{categoria_presupuestal}{chr. Código de programa presupuestal de forma "XXXX" (4).
Para Acciones Centrales usar "9001" y para APnoP usar "9002".}

\item{producto}{chr. Código de producto o proyecto de inversión de forma "XXXXXXX" (7).}

\item{actividad}{chr. Código de actividad, acción de inversión u obra, de forma "XXXXXXX" (7).}

\item{funcion}{chr. Código de función de forma "XX" (2).}

\item{division_funcional}{chr. Código de división funcional de forma "XXX" (3).}

\item{grupo_funcional}{chr. Código de grupo funcional de forma "XXXX" (4).}

\item{meta}{chr. Código de meta. No tiene forma estandarizada.}
}
\value{
Un data.frame con clase \if{html}{\out{<transpaeco>}}
}
\description{
Estas funciones permiten desglosar la busqueda de acuerdo al destino
de uso del presupuesto.
}
