% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elegir_institucion.R
\name{elegir_institucion}
\alias{elegir_institucion}
\alias{elegir_quien_gasta}
\alias{elegir_quien_recauda}
\title{Elegir la institucion que utiliza el presupuesto}
\usage{
elegir_institucion(
  x,
  nivel = NULL,
  sector = NULL,
  pliego = NULL,
  unidad_ejecutora = NULL,
  goblocal_o_manc = NULL,
  mancomunidad = NULL,
  departamento = NULL,
  provincia = NULL,
  municipalidad = NULL
)

elegir_quien_gasta(
  x,
  nivel = NULL,
  sector = NULL,
  pliego = NULL,
  unidad_ejecutora = NULL,
  goblocal_o_manc = NULL,
  mancomunidad = NULL,
  departamento = NULL,
  provincia = NULL,
  municipalidad = NULL
)

elegir_quien_recauda(
  x,
  nivel = NULL,
  sector = NULL,
  pliego = NULL,
  unidad_ejecutora = NULL,
  goblocal_o_manc = NULL,
  mancomunidad = NULL,
  departamento = NULL,
  provincia = NULL,
  municipalidad = NULL
)
}
\arguments{
\item{x}{Objeto de clase \verb{<transpaeco>} recipiente de los parámetros de la consulta deseada}

\item{nivel}{chr. Nivel de gobierno. Puede ser "E" (gobierno nacional),
"M" (gobiernos locales) o "R" (gobiernos regionales). Si es "E" o "R",
le corresponde \code{sector}. Si es "M", le corresponde \code{goblocal_o_manc}.}

\item{sector}{chr. Codigo de sector, de forma "XX" (2).}

\item{pliego}{chr. Código de pliego, de forma "XXX" (3).}

\item{unidad_ejecutora}{chr. Código de Unidad Ejecutora. No tiene forma estandar}

\item{goblocal_o_manc}{chr. Escoger entre "M" (gobierno local)
o "N" (mancomunidad). Si es "M", desbloque \code{departamento}, \code{provincia} y \code{municipalidad}.
Si es "N", le corresponde \code{mancomunidad}.}

\item{mancomunidad}{chr. Código de mancomunidad, de forma "XXXXXX" (6).}

\item{departamento}{chr. Código de departamento. El valor "07" corresponde a Callao.}

\item{provincia}{chr. Código de provincia. El valor "01" corresponde a capital de departamento.}

\item{municipalidad}{chr. Código de municipalidad, de forma "XXXXXX". No corresponde con ubigeo.}
}
\value{
Un data.frame con clase \verb{<transpaeco>}
}
\description{
Estas funciones permiten desglosar la busqueda haciendo uso del clasificador institucional.
\code{elegir_quien_gasta()} y \code{elegir_quien_recauda()} sirven los módulos de gasto e ingresos, respectivamente.
\code{elegir_institucion()} es una versión agnóstica que sirve para ámbos módulos.
}
