% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elegir_lugar.R
\name{elegir_lugar}
\alias{elegir_lugar}
\alias{elegir_donde_se_gasta}
\title{Elegir lugar de uso del presupuesto}
\usage{
elegir_lugar(x, departamento_meta = NULL)

elegir_donde_se_gasta(x, departamento_meta = NULL)
}
\arguments{
\item{x}{Objeto de clase \verb{<transpaeco>} recipiente de los parámetros de la consulta deseada}

\item{departamento_meta}{chr. Código de departamento.
Para obtener "Callao" usar "07", y para obtener 'Exterior' usar "98".}
}
\value{
Un data.frame con clase \if{html}{\out{<transpaeco>}}
}
\description{
Estas funciones permiten desglosar la busqueda de acuerdo
al lugar de uso del presupuesto.
}
