% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elegir_origen.R
\name{elegir_origen}
\alias{elegir_origen}
\alias{elegir_con_que_se_financia}
\alias{elegir_fuentes_de_recaudacion}
\title{Elegir el origen del presupuesto}
\usage{
elegir_origen(
  x,
  fuente_financiamiento = NULL,
  rubro = NULL,
  tipo_de_recurso = NULL
)

elegir_con_que_se_financia(
  x,
  fuente_financiamiento = NULL,
  rubro = NULL,
  tipo_de_recurso = NULL
)

elegir_fuentes_de_recaudacion(
  x,
  fuente_financiamiento = NULL,
  rubro = NULL,
  tipo_de_recurso = NULL
)
}
\arguments{
\item{x}{Objeto de clase \verb{<transpaeco>} recipiente de los parámetros de la consulta deseada}

\item{fuente_financiamiento}{int. Código de fuente de financiamiento.
En su versión más reciente va del 1 al 5.}

\item{rubro}{chr. código de rubro, de  forma "00" (2 digitos).}

\item{tipo_de_recurso}{chr, Código de tipo de recurso.
No tiene forma estandar, puede ser de una letra mayuscula,
un digito, o dos digitos.}
}
\value{
Un objeto con clase \verb{<transpaeco>}
}
\description{
Estas funciones permiten desglosar la busqueda de acuerdo al origen
del financiamiento del presupuesto.
\code{elegir_con_que_se_financia()} y \code{elegir_fuentes_de_recaudacion()} sirven los módulos de gasto e ingresos, respectivamente.
\code{elegir_origen()} es una versión agnóstica que sirve para ámbos módulos.
}
\examples{
iniciar_transparencia_economica(modulo = "gasto") \%>\%
    elegir_periodo_anual(2022) \%>\%
    elegir_con_que_se_financia(
        rubro = "todos"
    )

iniciar_transparencia_economica(modulo = "gasto") \%>\%
    elegir_periodo_anual(2022) \%>\%
    elegir_con_que_se_financia(
        rubro = "todos"
    ) \%>\%
    consultar()

# desglose de tipo de recurso del rubro de "13: Donaciones y transferencias"
# notese la variacion en los codigos de tipo de recurso
iniciar_transparencia_economica(modulo = "gasto") \%>\%
    elegir_periodo_anual(2022) \%>\%
    elegir_con_que_se_financia(
        rubro = "13",
        tipo_de_recurso = "todos"
    ) \%>\%
    consultar()

# cuando no se respeta la forma, el mensaje de error
# muestra la expresion regular esperada
tryCatch({
    iniciar_transparencia_economica(modulo = "gasto") \%>\%
        elegir_periodo_anual(2022) \%>\%
        elegir_con_que_se_financia(
            rubro = "13",
            tipo_de_recurso = "AA"
        )
}, error = function(e) print(e))

}
