% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elegir_years.R
\name{elegir_periodo_anual}
\alias{elegir_periodo_anual}
\title{Elegir periodo anual de consulta}
\usage{
elegir_periodo_anual(x, periodo)
}
\arguments{
\item{x}{Objeto de clase 'sep_df' recipiente de los parámetros de la consulta deseada}

\item{periodo}{numeric. Periodo anual de consulta}
}
\value{
sep_df
}
\description{
Elegir periodo anual de consulta
}
