% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elegir_tiempo.R
\name{elegir_tiempo}
\alias{elegir_tiempo}
\alias{elegir_cuando_se_hizo_gasto}
\alias{elegir_cuando_se_hizo_recaudacion}
\title{Elegir el tiempo en que se usa el presupuesto}
\usage{
elegir_tiempo(x, trimestre = NULL, mes = NULL)

elegir_cuando_se_hizo_gasto(x, trimestre = NULL, mes = NULL)

elegir_cuando_se_hizo_recaudacion(x, trimestre = NULL, mes = NULL)
}
\arguments{
\item{x}{Objeto de clase \verb{<transpaeco>} recipiente de los parámetros de la consulta deseada}

\item{trimestre}{int. Número del 1 al 4}

\item{mes}{int. Número del 1 al 12}
}
\value{
Un data.frame con clase \verb{<transpaeco>}
}
\description{
Estas funciones permiten desglosar la busqueda de acuerdo al tiempo en el periodo anual.
\code{elegir_cuando_se_hizo_gasto()} y \code{elegir_cuando_se_hizo_recaudacion()} sirven los módulos de gasto e ingresos, respectivamente.
\code{elegir_tiempo()} es una versión agnóstica que sirve para ámbos módulos.
}
