% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_html_body.R
\name{retrieve_html_body}
\alias{retrieve_html_body}
\title{Obtener el cuerpo de la respuesta HTML}
\usage{
retrieve_html_body(response)
}
\arguments{
\item{response}{Objeto de clase "httr2_response"}
}
\value{
Objeto de clase "html_node"
}
\description{
Esta operación toma la "httr2_response" definida y obtiene solo el cuerpo de la respuesta HTML. Ha sido definido así porque
sirve como paso intermedio para la obtención de la tabla.
}
